In the code above, we've added a new function called handleDeleteMessage to handle message deletion. When a client sends a request to delete a message, the server filters out the deleted message from the messages array and broadcasts the deleted message ID to all connected clients.
We've also added a new WebSocket event listener to handle incoming messages for message deletion. When a client sends a request to delete a message, the server calls the handleDeleteMessage function to remove the message from the messages array and broadcast the deleted message ID to all connected clients.01
In the code above, we've added a new function called handleUserDisconnection to handle user disconnection. When a client disconnects, the server removes the user from the activeUsers list.